requirements.txt
langchain==0.3.3
langchain_core==0.3.10
langchain_huggingface==0.1.0
python-dotenv==1.0.1
python-telegram-bot==21.6
